add discovery functionality

+++ src/db/models/event_settings_models.rs
@@ -12,7 +12,7 @@ pub struct EventSettings {
     pub event_types: Vec<String>, // Event types to subscribe to
     pub event_topic_expressions: Vec<String>, // ONVIF topic expressions
     pub trigger_recording: bool,  // Whether to trigger recording on events
-    pub recording_duration: u64,  // Duration to record in seconds when event triggered
+    pub recording_duration: i32,  // Duration to record in seconds when event triggered
     pub recording_quality: RecordingQuality, // Recording quality for event recordings
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
